+++ app/model/usdm/model/wrapper.py
@@ -1,0 +1,7 @@
+from usdm_model.wrapper import Wrapper
+from app.model.usdm.model.study import Study
+
+# def the_study(self: Wrapper) -> Study:
+#   return self.study
+
+# setattr(Wrapper, 'the_study', the_study)